// apps/service/src/common/core/core.enums.ts
export enum RepositoryEvents {
    BEFORE_LOAD = 'before_load',
    AFTER_LOAD = 'after_load',
    BEFORE_CREATE = 'befor_create',
    AFTER_CREATE = 'after_create',
    BEFORE_UPDATE = 'before_update',
    AFTER_UPDATE = 'after_update',
    BEFORE_REMOVE = 'before_remove',
    AFTER_REMOVE = 'after_remove'
}